src/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix>=11.21.0",
    "arize-phoenix-otel>=0.12.1",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.4",
    "openinference-instrumentation-langchain>=0.1.50",
]

[tool.uv]
package = true
